assignment

+++ twitoff/models.py
@@ -2,7 +2,7 @@ from flask_sqlalchemy import SQLAlchemy
 DB = SQLAlchemy()
 
 class User(DB.Model):
-    id = DB.Column(DB.Integer, primary_key=True)
+    id = DB.Column(DB.BigInteger, primary_key=True)
     name = DB.Column(DB.String(15), nullable=False)
     newest_tweet_id = DB.Column(DB.BigInteger)
 
@@ -15,5 +15,6 @@ class Tweet(DB.Model):
     embedding = DB.Column(DB.PickleType,nullable=False)
     user_id = DB.Column(DB.BigInteger, DB.ForeignKey('user.id'),nullable=False)
     user = DB.relationship('User',backref=DB.backref('tweets',lazy=True))
+    
     def __repr__(self):
         return '<Tweet {}>'.format(self.text)
